# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/v1/purchase-requests": {
            "get": {
                "tags": [
                    "Purchase Requests"
                ],
                "summary": "Get all purchase requests for the authenticated user",
                "operationId": "762973603ddae20d17be674612b9f7e7",
                "responses": {
                    "200": {
                        "description": "Purchase requests retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "purchase_requests": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Purchase Requests"
                ],
                "summary": "Create a purchase request",
                "operationId": "5f7dcb545d1177ffb645a2f1581fd147",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "needed_by"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laptop for project team"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Request for 2 business laptops"
                                    },
                                    "needed_by": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-10-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase request created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase-requests/{purchaseRequest}": {
            "get": {
                "tags": [
                    "Purchase Requests"
                ],
                "summary": "Get a single purchase request",
                "operationId": "c10a67408beee9c9860e5be757818d08",
                "parameters": [
                    {
                        "name": "purchaseRequest",
                        "in": "path",
                        "description": "Purchase request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase request retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "purchase_request": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Purchase Requests"
                ],
                "summary": "Update a purchase request",
                "operationId": "d80d3f19164e249c47bec18e2ed10330",
                "parameters": [
                    {
                        "name": "purchaseRequest",
                        "in": "path",
                        "description": "Purchase request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Laptop for project team"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated request for 3 laptops"
                                    },
                                    "needed_by": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-10-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase request updated successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Purchase Requests"
                ],
                "summary": "Delete a purchase request",
                "operationId": "d9e4baf475808465ea4773d2c1211e22",
                "parameters": [
                    {
                        "name": "purchaseRequest",
                        "in": "path",
                        "description": "Purchase request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase request deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase-requests/{purchaseRequest}/approve": {
            "patch": {
                "tags": [
                    "Purchase Requests"
                ],
                "summary": "Approve a pending purchase request",
                "operationId": "9d2bdf63cb2df4e2891dc62d4620468f",
                "parameters": [
                    {
                        "name": "purchaseRequest",
                        "in": "path",
                        "description": "Purchase request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "user created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user created successfully"
                                        },
                                        "member": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/purchase-requests/{purchaseRequest}/reject": {
            "patch": {
                "tags": [
                    "Purchase Requests"
                ],
                "summary": "Reject a pending purchase request",
                "operationId": "b9bc210030edf78e791fb8202bb7a365",
                "parameters": [
                    {
                        "name": "purchaseRequest",
                        "in": "path",
                        "description": "Purchase request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase request rejected successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}